separated tests for different apps & wrote more tests
