+++ django_imager/imager_profile/tests.py
@@ -3,10 +3,9 @@ from django.test import TestCase
 
 # Create your tests here.
 
-from imager_profile.models import ImagerProfile, User
+from imager_profile.models import User
 import factory
-from imager_images.models import Album, Photo
-from datetime import datetime
+from django.test import Client
 
 
 class UserFactory(factory.django.DjangoModelFactory):
@@ -20,7 +19,7 @@ class ProfileTests(TestCase):
 
     def setUp(self):
         """Initiate with two users in the db, one activce and one not."""
-        user = UserFactory.create()
+        user = UserFactory.build()
         user.username = 'john'
         user.email = 'john@image.com'
         user.set_password('password')
@@ -31,12 +30,8 @@ class ProfileTests(TestCase):
         user.profile.bio = 'I take pictures'
         user.profile.phone = '(000) 111 2222'
         user.profile.save()
-        album = Album(title='Album1', owner=user.profile)
-        album.save()  # to access this: user.profile.albums.first()
-        photo = Photo(title='Photo1', owner=user.profile)
-        photo.save()  # to access this: user.profile.photos.first()
 
-        user = UserFactory.create()
+        user = UserFactory.build()
         user.username = 'nick'
         user.email = 'nick@image.com'
         user.set_password('password')
@@ -47,10 +42,6 @@ class ProfileTests(TestCase):
         user.profile.bio = 'I take pictures too'
         user.profile.phone = '(000) 222 3333'
         user.profile.save()
-        album = Album(title='Album2', owner=user.profile)
-        album.save()  # to access this: user.profile.albums.first()
-        photo = Photo(title='Photo2', owner=user.profile)
-        photo.save()  # to access this: user.profile.photos.first()
 
     def test_user_nick_can_point_to_its_profile(self):
         """Test if user can point to its profile."""
@@ -120,62 +111,26 @@ class ProfileTests(TestCase):
         nick = User.objects.get(email='nick@image.com')
         self.assertIsNotNone(nick.profile.bio)
 
-    def test_user_john_has_album_in_profile(self):
-        """Test user have profile attribute album."""
-        john = User.objects.get(email='john@image.com')
-        self.assertIsNotNone(john.profile.albums.first())
 
-    def test_user_nick_has_album_in_profile(self):
-        """Test user have profile attribute album."""
-        nick = User.objects.get(email='nick@image.com')
-        self.assertIsNotNone(nick.profile.albums.first())
+class ViewTest(TestCase):
+    """Testing for the imager_profile app/model."""
 
-    def test_user_john_has_photo_in_profile(self):
-        """Test user have profile attribute photo."""
-        john = User.objects.get(email='john@image.com')
-        self.assertIsNotNone(john.profile.photos.first())
+    def setUp(self):
+        """Initiate with two users in the db, one activce and one not."""
+        self.client = Client()
 
-    def test_user_nick_has_photo_in_profile(self):
-        """Test user have profile attribute photo."""
-        nick = User.objects.get(email='nick@image.com')
-        self.assertIsNotNone(nick.profile.photos.first())
+    def test_can_access_main_view(self):
+        """Test home view is accessible."""
+        response = self.client.get('/')
+        self.assertEquals(response.status_code, 200)
 
-    def test_user_john_photo_has_create_date(self):
-        """Test photo has an upload date."""
-        john = User.objects.get(email='john@image.com')
-        self.assertIsInstance(john.profile.photos.first().date_uploaded, datetime)
+    def test_can_access_login_view(self):
+        """Test login view is accessible."""
+        response = self.client.get('/login')
+        self.assertEquals(response.status_code, 200)
 
-    def test_user_nick_photo_has_create_date(self):
-        """Test photo has an upload date."""
-        nick = User.objects.get(email='nick@image.com')
-        self.assertIsInstance(nick.profile.photos.first().date_uploaded, datetime)
+    def test_can_access_logout_view(self):
+        """Test logout view is accessible."""
+        response = self.client.get('/logout')
+        self.assertEquals(response.status_code, 200)
 
-    def test_user_john_photo_has_publised_type(self):
-        """Test photo has an published type."""
-        john = User.objects.get(email='john@image.com')
-        self.assertEquals(john.profile.photos.first().published, 'PRIVATE')
-
-    def test_user_nick_photo_has_published_type(self):
-        """Test photo has an published type."""
-        nick = User.objects.get(email='nick@image.com')
-        self.assertEquals(nick.profile.photos.first().published, 'PRIVATE')
-
-    def test_user_john_album_has_create_date(self):
-        """Test album has an upload date."""
-        john = User.objects.get(email='john@image.com')
-        self.assertIsInstance(john.profile.albums.first().date_uploaded, datetime)
-
-    def test_user_nick_album_has_create_date(self):
-        """Test album has an upload date."""
-        nick = User.objects.get(email='nick@image.com')
-        self.assertIsInstance(nick.profile.albums.first().date_uploaded, datetime)
-
-    def test_user_john_album_has_publised_type(self):
-        """Test album has an published type."""
-        john = User.objects.get(email='john@image.com')
-        self.assertEquals(john.profile.albums.first().published, 'PRIVATE')
-
-    def test_user_nick_album_has_published_type(self):
-        """Test album has an published type."""
-        nick = User.objects.get(email='nick@image.com')
-        self.assertEquals(nick.profile.albums.first().published, 'PRIVATE')
