+++ django_imager/imager_profile/urls.py
@@ -3,6 +3,6 @@ from django.conf.urls import url, include
 from django.contrib.auth.views import login, logout
 
 urlpatterns = [
-    url(r'^registration/', include("registration.backends.simple.urls")),
+    url(r'^registration', include("registration.backends.simple.urls")),
 
 ]
